+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Fixes
 
 * [#1570](https://github.com/ruby-grape/grape/pull/1570): Make versioner consider the mount destination path - [@namusyaka](https://github.com/namusyaka).
+* [#1579](https://github.com/ruby-grape/grape/pull/1579): Fix delete status with a return value - [@eproulx-petalmd](https://github.com/eproulx-petalmd).
 * Your contribution here.
 
 #### Features